LSP: Move LSPManager class to editor/lsp/manager.py
